api/requirements.txt
fastapi==0.104.0
pandas==2.0.3
python-dotenv==1.0.0
scikit_learn==1.2.1
SQLAlchemy==2.0.6
uvicorn==0.23.2